agent-service/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
boto3>=1.34.0
aiohttp>=3.9.0
numpy>=1.24.0
python-multipart>=0.0.6
strands-agents>=1.0.0
